model/requirements.txt
tensorflow[and-cuda]==2.14.0
keras==2.14.0
scikit-learn
numpy
huggingface_hub
opencv-python
torch
pillow
